app/requirements.txt
aerich==0.5.8
aiohttp==3.7.4.post0
aiosqlite==0.17.0
asgiref==3.4.1
async-timeout==3.0.1
asyncpg==0.24.0
attrs==21.2.0
cffi==1.14.6
chardet==4.0.0
click==8.0.1
cryptography==3.4.8
ddlparse==1.10.0
dictdiffer==0.9.0
fastapi==0.68.1
gunicorn==20.1.0
h11==0.12.0
idna==3.2
iso8601==0.1.16
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==5.1.0
passlib==1.7.4
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pyparsing==2.4.7
pypika-tortoise==0.1.1
pytz==2021.1
starlette==0.14.2
tortoise-orm==0.17.7
typing-extensions==3.10.0.2
uvicorn==0.15.0
xmltodict==0.12.0
yarl==1.6.3
